[!!!][TASK] Move token expiration calulation to ResourceTarget
Previously the token only contained details about the *creation*
timestamp. Expiriration of a token was calculated in the HTTP
component.

With this change the token actually contains the expiration date
allowing for:

* specifying multiple targets with different lifetimes
* chaning the lifetime without affecting published tokens

This is a breaking change because it moves the "tokenLifetime"
setting from
``TYPO3.Flow.http.chain.process.chain.protectedResources.componentOptions``
to
``TYPO3.Flow.resource.targets.protectedResourcesTarget.targetOptions``

and you'll have to adjust your settings, in case you used the "tokenLifetime"
option.