!!!TASK: Use $resource in FileServeStrategyInterface
Currently the `$filePathAndName` is used directly in the method
`serve()`. This prevents the development of strategies to deliver files
that are not identified by file names.

This breaking change replaces `$filePathAndName` with `$resource`.
If a strategy is used to serve files from `FileSystemStorage` it can
use the new static method
`ProtectedResourceUtility:getStoragePathAndFilenameByHash()` to find the
local file name.